Restructure Udger client project
